+++ libs/naas-abi-core/naas_abi_core/module/ModulePipelineLoader.py
@@ -1,0 +1,8 @@
+from naas_abi_core.module.ModuleComponentLoader import load_subclasses
+from naas_abi_core.pipeline.pipeline import Pipeline
+
+
+class ModulePipelineLoader:
+    @classmethod
+    def load_pipelines(cls, class_: type) -> list[type[Pipeline]]:
+        return load_subclasses(class_, "pipelines", Pipeline)